requirements.txt
blinker==1.8.2
click==8.1.7
Flask==2.0.3
importlib_metadata==8.2.0
itsdangerous==2.0.0
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.3
zipp==3.6.0
psycopg2-binary==2.9.1
